+++ nix/package.nix
@@ -13,7 +13,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "eilmeldung";
-  version = "0.4.4";
+  version = "0.4.5";
   
   src = ../.;